requirements.txt
datasets==2.10.1
Flask==2.2.3
Flask_RESTful==0.3.9
ipdb==0.13.9
numpy==1.24.3
openprompt==1.0.1
pandarallel==1.6.3
pandas==1.3.4
redditcleaner==1.1.2
scikit_learn==0.24.2
torch==1.12.1+cu113
transformers==4.28.1
tweet_preprocessor==0.6.0
